Allow warnings when compiling documentation
Problem

To build the documentation on readthedocs, we use Sphinx, additionally with the
-W flag, which elevates all warnings to errors. For some reason, Sphinx warns
when compiling pysmurf.core.devices._UdpReceiver, which elevates to error,
which prevents the remote documentation tests from finishing.

Solution

Allow warnings when compiling the documentation. We can deal with the warnings
later. Coming from experience, Sphinx loves to error on many, many things, and
fixing all of Sphinx's warnings is usually not a good use of time.