Recreate Rogue v4.11.8 image to look for error
I built the smurf-rogue-docker image using Rogue v4.11.9, and then plugged it
into pysmurf, which then failed to import rogue. I downgraded pysmurf to an old
version of smurf-rogue-docker to check that it can run its pysmurf tests, which
passed. Now I'll recreate the smurf-rogue-docker image again using Rogue
v4.11.8 to check if the image if the problem is with the image builing process.
If this commit fails to build, then the issue is with the image building
process in the smurf-rogue-docker repo. If this passes, then the issue is with
the Rogue repository and how we're compiling it.