Flux ramp synchronized to fiber timing or backplane timing
This change permits use of the SMuRF timing crate for receiving timing from the
fiber or backplane for the purpose of synchronizing the flux ramp of multiple
SMuRF carriers.

This code assumes three use cases for SMuRF carriers: 1. their timing is
internal, 2. their timing is received by the timing fiber connected to its RTM,
3. their timing is received by the backplane of its SMuRF carrier. To support
these three use cases, the pysmurf cfg file was changed to support three
configurations into the "timing_reference" key: "ext_ref", "fiber", and
"backplane". Configuration on ext_ref is unchanged. Configuration on "fiber"
now receives the timing signal from the fiber connected on the RTM, and
distributes that signal to the rest of the backplane. Configurtion on backplane
will receive these triggers and lock its flux ramp to that signal.