Add control of the second 50K on the C04
This code change permits use of the second 50K Low Noise Amplifier
(LNA) on the C04 cryocard. This code should be backwards compatible
with the existing users on the C02 cryocard. The config file
cfg_files/rflab/rflab.cfg was used to test the functionality according
to the C02 testing procedure with the new functions. More thorough
testing that was not done is a. measure resistor values, b. take data
from the cryostat using the C04 and this code.