Add function to enable the LMK
Problem

The AMC LMK is needed to process the fiber or backplane timing. The
registers 0x146 and 0x147 need to be set dynamically depending on the
setup. However, setting the value does nothing if LMK:enable is 0.

Solution

Add the enable variable for the LMK to allow modifying it on the
fly. This is yet another variable that should be moved to the Rogue
init layer. Someone should think more carefully about how to do this.