Control flux ramp AC/DC coupling on the C02 and C04 Cryocard
Problem

Pysmurf should be able to control the flux ramp coupling mode for the
C02 and C04, either DC mode or AC mode. To do this, the C02 controls
two relays, K1 and K2, while the C04 uses just one relay named K2.

Solution

The easiest solution is to set bit 16 on address 0x2 on the PIC. This
controls both relays on the C02, and just the K2 on the C04, which is
very convenient. To read back the AC/DC status, just read back the
same bit. This deprecates the usage of address 0x08 on the C02
firmware. If we still want to use that address, the code gets more
complicated.

Tests

Tested using C04 firmware R4.0.7 and C02 firmware R1.1.0.