docs: pin sphinx version for rtd build (#702)
This adds a requirements file for the docs builds on readthedocs. In this, we pin the Sphinx version, which resolves the issue seen in https://github.com/slaclab/pysmurf/issues/701. Note build.tools.python fails on build https://readthedocs.org/projects/pysmurf/builds/15917436/ for some reason, so python.version is used for now.

Thanks @BrianJKoopman.