Fix issue where AMC LMK doesn't lock to the FPGA clock (#705)
Problem

The AMC LMK clock receives 122.88 MHz, cleans it up, then spits it
back out to the FPGA. To do this, it locks to one of two clocks,
CLKin0 or CLKin1. 0 is the FPGA's recovered timing clock, and 1 is the
front panel clock. Previously, with fiber timing, the LMK still tries
to lock to the front panel, fails, then free runs, preventing the
tones from locking to the fiber 122.88 MHz.

Solution

Set LMK 0x146 to 0x08, CLKin0, when taking timing from fiber. This also
applies to the backplane case, so set that as well.

Problems

This has to be dynamically set after Rogue's init, because pysmurf
intentionally resets the RTM, which resets any parameters set on
the LMK. We should think more carefully about how to minimize
the steps of the SMuRF boot process.