Add control for the Cryocard C04 50K2 amplifier (#697)
This adds control for the 50K2 amplifier on the C04 cryocard. No changes are
made to the other three amplifiers, or TES biases. For the 50K2, the gate and
drain are set by RTM DAC. The get and set functions for the 50K2 gate and drain
voltage are relative to what goes out of the cryocard, not the voltage from the
RTM DAC to the cryocard. This means the get fucntions are not true measurements
of those voltages but approximations made from converting the known DAC voltage
with the parameters fiftyk2_gate_bit_to_V, fiftyk2_drain_conversion_m, and
fiftyk2_drain_conversion_b.